Initial commit
I'll probably have to edit a few things with the suggestions of some of
the teammates, but this is overall the finished product. While this is
supposed to be article-like, I felt that some parts were a little "wall
of text"-ish, so that might be something I come back to.